Naming Changes / Fixes from Review
