+++ src/main/java/frc/robot/commands/shooter/SetShooterFromTriggerDebug.java
@@ -23,7 +23,7 @@ public class SetShooterFromTriggerDebug extends CommandBase {
 
     @Override
     public void end(boolean interrupted) {
-        flywheelSubsystem.stop();
+        flywheelSubsystem.stopFlywheel();
     }
 
     @Override
